feat: при активном фильтре вокруг кнопки появляется градиентная граница градиентная граница вынесена из StatisticsCell в отдельный файл
